refactor(placement): move the bridge repair pass into placement
`_reconcile_bridge_attachments` and `_record_bridge_network` move out of
`deploy_manager` into `placement` as `repair` and `record_bridge_network`.
Both are called from another module, so neither keeps its underscore.

`ingress.reconcile` calls `placement.repair()` through a module-level
import. The lazy import pass 1 left behind is deleted: `ingress` imports
`placement` and `placement` imports neither, so there is no cycle left to
work around.

`repair` now reports what it found as well as what it fixed. Without the
split, a bridge nothing could be reattached to returns exactly what a
healthy bridge returns.

`TestReconcileBridgeAttachments` moves to `test_placement.py` as `TestRepair`.
