+++ benchpress/tests/test_deploy_manager.py
@@ -10,12 +10,11 @@ from pathlib import Path
 from types import SimpleNamespace
 from unittest.mock import MagicMock, patch
 
-import docker
 import frappe
 import yaml
 from frappe.tests import IntegrationTestCase
 
-from benchpress import ingress, placement
+from benchpress import ingress
 from benchpress.benchpress.doctype.bench_instance import get_instance_id
 from benchpress.tests.test_docker_manager import exec_commands, exec_environments
 
@@ -1403,61 +1402,3 @@ class TestRecordPrimarySite(IntegrationTestCase):
 		with tempfile.TemporaryDirectory() as tmp:
 			with patch.object(ingress, "TRAEFIK_DYNAMIC_DIR", _mounted(tmp)):
 				teardown_bench(bench)
-
-
-class TestReconcileBridgeAttachments(IntegrationTestCase):
-	"""The half of the `*/5` pass that owns which containers can reach a bench bridge.
-
-	`docker compose up -d traefik` recreates the proxy with only its compose networks, so
-	every bridge this app made loses its ingress and says nothing about it.
-	"""
-
-	def _reconcile(self, present: dict[str, list[str]], bridges=2):
-		"""Run the pass against a daemon whose bridges hold `present`; returns (result, connects)."""
-		from benchpress import deploy_manager
-
-		client = MagicMock()
-		connects = []
-
-		def get(name):
-			if name not in present:
-				raise docker.errors.NotFound(name)
-			network = MagicMock()
-			network.attrs = {"Containers": {n: {"Name": n} for n in present[name]}}
-			network.connect.side_effect = lambda container: connects.append((name, container))
-			return network
-
-		client.networks.get.side_effect = get
-		with (
-			patch("benchpress.docker_manager.get_client", return_value=client),
-			patch("benchpress.placement.bridge_count", return_value=bridges),
-		):
-			return deploy_manager._reconcile_bridge_attachments(), connects
-
-	def test_a_recreated_proxy_is_put_back_without_being_restarted(self):
-		result, connects = self._reconcile(
-			{"benchpress-0": ["benchpress-mariadb", "benchpress-redis", "a-bench"]}
-		)
-
-		self.assertEqual(connects, [("benchpress-0", "benchpress_traefik")])
-		self.assertEqual(result, {"attached": {"benchpress-0": ["benchpress_traefik"]}})
-
-	def test_a_quiet_tick_reports_nothing(self):
-		result, connects = self._reconcile({"benchpress-0": list(placement.INFRASTRUCTURE_CONTAINERS)})
-
-		self.assertEqual(connects, [])
-		self.assertEqual(result, {"attached": {}})
-
-	def test_a_bridge_that_does_not_exist_is_not_created(self):
-		"""The family grows on a deploy, never on a timer."""
-		_result, connects = self._reconcile({})
-
-		self.assertEqual(connects, [])
-
-	def test_only_the_three_infrastructure_containers_are_ever_connected(self):
-		_result, connects = self._reconcile({"benchpress-0": [], "benchpress-1": []})
-
-		self.assertEqual(
-			{container for _network, container in connects},
-			set(placement.INFRASTRUCTURE_CONTAINERS),
-		)
